Move MessageSource implementations from `.message` to `.message.source`
