+++ mirai-core/src/commonMain/kotlin/message/source/incomingSourceImpl.kt
@@ -9,7 +9,7 @@
 
 @file:Suppress("EXPERIMENTAL_API_USAGE", "EXPERIMENTAL_OVERRIDE", "INVISIBLE_REFERENCE", "INVISIBLE_MEMBER")
 
-package net.mamoe.mirai.internal.message
+package net.mamoe.mirai.internal.message.source
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
@@ -22,6 +22,9 @@ import net.mamoe.mirai.internal.contact.GroupImpl
 import net.mamoe.mirai.internal.contact.checkIsGroupImpl
 import net.mamoe.mirai.internal.contact.newAnonymous
 import net.mamoe.mirai.internal.getGroupByUinOrCodeOrFail
+import net.mamoe.mirai.internal.message.IncomingMessageSourceInternal
+import net.mamoe.mirai.internal.message.MessageSourceSerializerImpl
+import net.mamoe.mirai.internal.message.toMessageChainNoSource
 import net.mamoe.mirai.internal.network.protocol.data.proto.ImMsgBody
 import net.mamoe.mirai.internal.network.protocol.data.proto.MsgComm
 import net.mamoe.mirai.internal.network.protocol.data.proto.SourceMsg
