+++ mirai-core/src/commonMain/kotlin/message/source/offlineSourceImpl.kt
@@ -8,11 +8,15 @@
  */
 
 
-package net.mamoe.mirai.internal.message
+package net.mamoe.mirai.internal.message.source
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import net.mamoe.mirai.Bot
+import net.mamoe.mirai.internal.message.MessageSourceInternal
+import net.mamoe.mirai.internal.message.MessageSourceSerializerImpl
+import net.mamoe.mirai.internal.message.toMessageChainNoSource
+import net.mamoe.mirai.internal.message.toRichTextElems
 import net.mamoe.mirai.internal.network.protocol.data.proto.ImMsgBody
 import net.mamoe.mirai.internal.network.protocol.data.proto.MsgComm
 import net.mamoe.mirai.internal.network.protocol.data.proto.SourceMsg
